Fixed audio playback after level
GameObject.Instantiate clones the GameObject and puts it in the root of the scene. We have to setparent it to our viewcontroller so it is disabled. It also had a visual component to it (the moving ball in the normal latency adjuster) so I had to disable it in the constructor.